@@ -358,7 +358,7 @@ class Context:
 
     def fill(self):
         self._xy = (0, 0)
-        raise NotImplementedError()
+        raise NotImplementedError
 
     def set_font(self, style, bold=False, italic=False):
         if style not in ("serif", "sans-serif", "monospaced"):
@@ -887,7 +887,7 @@ class LBRN2Surface(Surface):
                             C, x, y = c[0:3]
                             if C == "M":
                                 if start and points_equal(start[1], start[2], x0, y0):
-                                    pl.text = "LineClosed"
+                                    pl.text += f"L{cnt-1} 0"
                                 start = c
                                 cnt = 1
                                 if self.dbg: print ("next, because M")
@@ -915,6 +915,8 @@ class LBRN2Surface(Surface):
                         if start and points_equal(start[1], start[2], x0, y0):
                                 if bspline == False:
                                     pl.text = "LineClosed"
+                                else:
+                                    pl.text += f"L{cnt-1} 0"
                         start = c
                         if self.dbg: print ("2", num)
                     elif C == "T":
