@@ -48,11 +48,11 @@ class AllEdges(Boxes):
             with self.saved_context():
                 self.move(0, 0, "", True)
                 self.moveTo(x, 0, 90)
-                self.edge(t+self.edges[c].startwidth())
+                self.edge(t+self.edges[c].startWidth())
                 self.corner(90)
                 self.edges[c](x, h=4*t)
                 self.corner(90)
-                self.edge(t+self.edges[c].endwidth())
+                self.edge(t+self.edges[c].endWidth())
                 self.move(0, 0, "")
 
             self.moveTo(0, 3*t + self.edges[c].spacing())
