@@ -91,8 +91,8 @@ class DiceBox(Boxes):
 
         t = self.thickness
 
-        hy = self.edges["O"].startwidth()
-        hy2 = self.edges["P"].startwidth()
+        hy = self.edges["O"].startWidth()
+        hy2 = self.edges["P"].startWidth()
 
         e1 = edges.CompoundEdge(self, "eF", (hy-t, h-hy+t))
         e2 = edges.CompoundEdge(self, "Fe", (h-hy+t, hy-t))
