@@ -10,6 +10,7 @@ import gettext
 import os
 import sys
 import argparse
+import logging
 from pathlib import Path
 
 try:
@@ -128,14 +129,15 @@ def example_output_fname_formatter(box_type, name, box_idx, metadata, box_args):
 
 
 def main() -> None:
-    parser = argparse.ArgumentParser(formatter_class=argparse.RawDescriptionHelpFormatter, description=__doc__)
+    parser = argparse.ArgumentParser(formatter_class=argparse.RawDescriptionHelpFormatter, description=__doc__, add_help=False)
     parser.allow_abbrev = False
     parser.add_argument("--generator", type=str, default=None)
     parser.add_argument("--id", type=str, default=None, help="ignored")
-    parser.add_argument("--debug", action="store_true", default=False)
+    parser.add_argument("--debug", type=boxes.boolarg, default=False)
     parser.add_argument("--version", action="store_true", default=False)
     parser.add_argument("--list", action="store_true", default=False, help="List available generators.")
     parser.add_argument("--examples", action="store_true", default=False, help='Generates an SVG for every generator into the "examples" folder.')
+    parser.add_argument("--help", action="store_true", default=False)
     args, extra = parser.parse_known_args()
     if args.generator and (args.examples or args.list):
         parser.error("cannot combine --generator with other commands")
@@ -154,13 +156,19 @@ def main() -> None:
     else:
         if args.generator:
             name = args.generator
-        else:
+        elif extra:
             name = extra.pop(0).lower()
+        else:
+            parser.print_help()
+            sys.exit(0)
+        if args.help:
+            extra.append("--help")
+        if args.debug:
+            extra.extend(["--debug", "1"])
         run_generator(name, extra)
 
 if __name__ == '__main__':
     # Setup basic logging
-    import logging
     logging.basicConfig(level=logging.INFO)
 
     main()
